requirements.txt
Django==1.9.4
django-ogmios==0.9.3
django-authtools==1.4.0
djangorestframework==3.3.3
requests==2.9.1
django-argonauts==1.1.4
python-gnupg==0.3.8
PyJWT==1.4.0
env-excavator==1.5.0
python-dotenv==0.1.0
dj-database-url==0.4.0
psycopg2==2.6.1
gunicorn==19.4.5
django-storages==1.4.1
boto==2.39.0
django-s3-folder-storage==0.3
jsonschema==2.5.1
ipfs-api==0.2.3
django-tables2==1.2.0
django-materializecss-form==1.0.1
djangorestframework-filters==0.8.0
